move get_model_from_config
